Define unique pkg-config search path for OpenEXR
On MacOS, the brew version of OpenEXR may not compile (I can't rememer the problem and didn't capture it). Anyway, it may be necessary to build from source which may require a unique location of your pkg-config .pc files. This enhancement, will search the provided library path to extrapolate the pkg-config location and define the CPPFLAGS and LDFLAGS paths.