Refrain from issuing an error when `#version 3.8` seems to be specified late.
Previous v3.8 development versions issued an outright error when the scene did not start with `#version`, but specified `#version 3.8` or higher later; the rationale was that such scenes would clearly not be legacy scenes, and the lack of a now-mandatory leading `#version` statement therefore safe to be considered an outright error without breaking backward compatibility. However, the current implementation fails to account for a fringe scenario involving the `Include_Header` INI setting. We're therefore disabling this error for now, and just let the parser issue a warning as in v3.7.